Updated PR comments and added documentation
