+++ Sources/ETCDExample/ETCDExample.swift
@@ -18,52 +18,56 @@ import Dispatch
 @main
 struct Example {
     static func main() async throws {
-        let eventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 1)
+        let eventLoopGroup = MultiThreadedEventLoopGroup.singleton
         do {
             let etcdClient = EtcdClient(host: "localhost", port: 2379, eventLoopGroup: eventLoopGroup)
-            Task {
-                do {
-                    try await etcdClient.watch("foo") { sequence in
-                        var iterator = sequence.makeAsyncIterator()
-                        while let event = try await iterator.next() {
-                            print(event)
+            try await withThrowingTaskGroup(of: Void.self) { group in
+                group.addTask {
+                    do {
+                        try await etcdClient.watch("foo") { sequence in
+                            var iterator = sequence.makeAsyncIterator()
+                            while let event = try await iterator.next() {
+                                print(event)
+                            }
+                        }
+                    } catch {
+                        print("Error watching key: \(error)")
+                    }
+                }
+                // Sleeping for a second to let the watch above setup
+                try await Task.sleep(for: .seconds(1))
+                
+                try await etcdClient.set("foo", value: "bar")
+                if let value = try await etcdClient.get("foo") {
+                    if let stringValue = String(data: value, encoding: .utf8) {
+                        print("Value is: \(stringValue)")
+                        try await etcdClient.delete("foo")
+                        print("Key deleted")
+                        
+                        // Trying to get the value again
+                        let deletedValue = try await etcdClient.get("foo")
+                        if deletedValue == nil {
+                            print("Key not found after deletion")
+                        } else {
+                            print("Value after deletion: \(deletedValue!)")
+                        }
+                    } else {
+                        print("Unable to get value")
+                    }
+                } else {
+                    print("Key not found")
+                }
+                DispatchQueue.global().asyncAfter(deadline: .now() + 5) {
+                    Task {
+                        do {
+                            try await etcdClient.set("foo", value: "updated_value")
+                        } catch {
+                            print("Error setting updated value: \(error)")
                         }
                     }
-                } catch {
-                    print("Error watching key: \(error)")
                 }
+                try await eventLoopGroup.next().makePromise(of: Void.self).futureResult.get()
             }
-            
-            try await etcdClient.set("foo", value: "bar")
-            if let value = try await etcdClient.get("foo") {
-               if let stringValue = String(data: value, encoding: .utf8) {
-                   print("Value is: \(stringValue)")
-                   try await etcdClient.delete("foo")
-                   print("Key deleted")
-                   
-                   // Trying to get the value again
-                   let deletedValue = try await etcdClient.get("foo")
-                   if deletedValue == nil {
-                       print("Key not found after deletion")
-                   } else {
-                       print("Value after deletion: \(deletedValue!)")
-                   }
-               } else {
-                   print("Unable to get value")
-               }
-            } else {
-               print("Key not found")
-            }
-            DispatchQueue.global().asyncAfter(deadline: .now() + 5) {
-                Task {
-                    do {
-                        try await etcdClient.set("foo", value: "updated_value")
-                    } catch {
-                        print("Error setting updated value: \(error)")
-                    }
-                }
-            }
-            try await eventLoopGroup.next().makePromise(of: Void.self).futureResult.get()
         } catch {
             print("Error: \(error)")
         }
